+++ fillit_dir/fillit.h
@@ -17,25 +17,29 @@
 # include <fcntl.h>
 # include <libft.h>
 
-struct			s_tetrimino
+struct					s_tetrimino
 {
-	uint64_t	binary_tetrimino;
-	short int	cubes[5][2];
-	short int	placed;
+	unsigned long long	binary_tetrimino;
+	short int			cubes[5][2];
+	short int			placed;
 };
 
-int				opening(char *str, struct s_tetrimino tetriminos[27]);
-void			cubes_offsetter(uint64_t tetrimino, short int cubes[5][2]);
-short int		height_initializer(struct s_tetrimino *tetriminos,
+int						opening(char *str, struct s_tetrimino tetriminos[27]);
+void					cubes_offsetter(unsigned long long tetrimino,
+	short int cubes[5][2]);
+short int				height_initializer(struct s_tetrimino *tetriminos,
 	short int i, short int size);
-short int		width_initializer(struct s_tetrimino *tetriminos,
+short int				width_initializer(struct s_tetrimino *tetriminos,
 	short int i, short int size, short int *w_initialized);
-void			place_tetri(struct s_tetrimino *tetrimino, uint64_t *grid);
-void			remove_tetri(struct s_tetrimino *tetriminos, uint64_t *grid);
-void			algorithm(struct s_tetrimino *tetriminos,
-	uint64_t *grid, short int *size);
-short int		next_unplaced(struct s_tetrimino *tetriminos, short int i);
-uint64_t		shift(uint64_t tetro);
-void			total_read(int i, int bytes_read);
+void					place_tetri(struct s_tetrimino *tetrimino,
+	unsigned long long *grid);
+void					remove_tetri(struct s_tetrimino *tetriminos,
+	unsigned long long *grid);
+void					algorithm(struct s_tetrimino *tetriminos,
+	unsigned long long *grid, short int *size);
+short int				next_unplaced(struct s_tetrimino *tetriminos,
+	short int i);
+unsigned long long		shift(unsigned long long tetro);
+void					total_read(int i, int bytes_read);
 
 #endif
